+++ src/layers/ClientCluster.js
@@ -18,11 +18,14 @@ class ClientCluster extends Cluster {
             strokeColor = eventStrokeColor,
             countColor = clusterCountColor,
         } = this.options
+        const isInteractive = true
 
         super.createLayers()
 
         // Clusters
-        this.addLayer(clusterLayer({ id, color, strokeColor }), true)
+        this.addLayer(clusterLayer({ id, color, strokeColor }), {
+            isInteractive,
+        })
         this.addLayer(clusterCountLayer({ id, color: countColor }))
     }
 
@@ -45,8 +48,10 @@ class ClientCluster extends Cluster {
         if (this._hasPolygons) {
             const mapgl = this.getMapGL()
 
-            mapgl.off('sourcedata', this.onSourceData)
-            mapgl.off('moveend', this.updatePolygons)
+            if (mapgl) {
+                mapgl.off('sourcedata', this.onSourceData)
+                mapgl.off('moveend', this.updatePolygons)
+            }
         }
     }