+++ src/layers/__tests__/EarthEngine.spec.js
@@ -1,4 +1,4 @@
-import EarthEngine, { defaultOptions } from '../EarthEngine'
+import EarthEngine from '../EarthEngine'
 
 const urlFormat =
     'https://earthengine.googleapis.com/v1alpha/projects/earthengine-legacy/maps/.../tiles/{z}/{x}/{y}'
@@ -172,14 +172,5 @@ describe('EarthEngine', () => {
         const features = layer.getFeatures()
 
         expect(features.some(f => f.geometry.type === 'Point')).toBe(false)
-    })
-
-    it('Should filter features based on ids', async () => {
-        const layer = new EarthEngine(options)
-
-        expect(layer.getFilteredFeatures().length).toBe(3)
-        expect(layer.getFilteredFeatures([]).length).toBe(0)
-        expect(layer.getFilteredFeatures(['O6uvpzGd5pu']).length).toBe(1)
-        expect(layer.getFilteredFeatures(['O6uvpzGd5pu', 'abc']).length).toBe(1)
     })
 })