+++ src/utils/core.js
@@ -1,0 +1,3 @@
+// Replaces {key} with data in a string template
+export const setTemplate = (text, data) =>
+    text.replace(/\{ *([\w_-]+) *\}/g, (str, key) => data[key])