+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## [3.0.9](https://github.com/dhis2/maps-gl/compare/v3.0.8...v3.0.9) (2022-02-15)
+
+
+### Bug Fixes
+
+* check if map still exists before removing layer events ([#445](https://github.com/dhis2/maps-gl/issues/445)) ([c97384f](https://github.com/dhis2/maps-gl/commit/c97384f9ddc8d8481a7b02777d0f2f904d54b86a))
+
+## [3.0.8](https://github.com/dhis2/maps-gl/compare/v3.0.7...v3.0.8) (2022-02-13)
+
+
+### Bug Fixes
+
+* check if map still exists before collapsing a cluster (DHIS2-12583) ([#443](https://github.com/dhis2/maps-gl/issues/443)) ([46d2f15](https://github.com/dhis2/maps-gl/commit/46d2f154df1ccbfdf66bbce493796ec9be72de4a))
+
+## [3.0.7](https://github.com/dhis2/maps-gl/compare/v3.0.6...v3.0.7) (2022-02-09)
+
+
+### Bug Fixes
+
+* hide label on mouse move ([#442](https://github.com/dhis2/maps-gl/issues/442)) ([d1f14c9](https://github.com/dhis2/maps-gl/commit/d1f14c9a67852d8f701b8f6d11a10058f860c707))
+
+## [3.0.6](https://github.com/dhis2/maps-gl/compare/v3.0.5...v3.0.6) (2022-02-08)
+
+
+### Bug Fixes
+
+* redraw layer group on basemap change ([#438](https://github.com/dhis2/maps-gl/issues/438)) ([03aff61](https://github.com/dhis2/maps-gl/commit/03aff61e327d56d000e1547540ae0f95b598dc87))
+
+## [3.0.5](https://github.com/dhis2/maps-gl/compare/v3.0.4...v3.0.5) (2022-02-07)
+
+
+### Bug Fixes
+
+* make sure map is present before removing layers and source ([#436](https://github.com/dhis2/maps-gl/issues/436)) ([58c6cd1](https://github.com/dhis2/maps-gl/commit/58c6cd1c01267294163cabcb8206e528a8a18ef3))
+
+## [3.0.4](https://github.com/dhis2/maps-gl/compare/v3.0.3...v3.0.4) (2022-01-25)
+
+
+### Bug Fixes
+
+* handle ee null values ([#433](https://github.com/dhis2/maps-gl/issues/433)) ([3cd806b](https://github.com/dhis2/maps-gl/commit/3cd806b2f0010025f65086ce477c9841c25d1bd9))
+
+## [3.0.3](https://github.com/dhis2/maps-gl/compare/v3.0.2...v3.0.3) (2022-01-24)
+
+
+### Bug Fixes
+
+* keep data filter when earth engine layer is redrawn ([#431](https://github.com/dhis2/maps-gl/issues/431)) ([b74c12f](https://github.com/dhis2/maps-gl/commit/b74c12fe9556a746ad3e1e0a7095704bab83499e))
+* removed ee layer unit test ([0574148](https://github.com/dhis2/maps-gl/commit/057414857279b13f1487428bd12378265a5de6de))
+
 ## [3.0.2](https://github.com/dhis2/maps-gl/compare/v3.0.1...v3.0.2) (2022-01-04)